backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.20
alembic>=1.13.0
# psycopg2-binary is only needed for PostgreSQL (Docker)
# For local dev, SQLite is used automatically

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Task Queue (optional for local dev)
celery[redis]>=5.3.0
redis>=5.0.0

# OCR
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0

# ML & Data Science
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0

# Image Processing & Forgery Detection
opencv-python-headless>=4.9.0
imagehash>=4.3.0

# Fuzzy Matching
jellyfish>=1.0.0
rapidfuzz>=3.6.0

# Storage (MinIO S3)
minio>=7.2.0

# Encryption
cryptography>=42.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# WebSocket
websockets>=12.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
